+++ demo/chat/src/constants/routes.jsx
@@ -1,0 +1,14 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+const awsPath = '/Prod';
+export const rootPath = window.location.href.includes(awsPath)
+  ? `${awsPath}/`
+  : '/';
+
+const routes = {
+  SIGNIN: `${rootPath}`,
+  CHAT: `${rootPath}rooms`,
+};
+
+export default routes;