+++ src/components/ui/PopOver/Styled.tsx
@@ -7,7 +7,6 @@ import { ellipsis } from '../../../utils/style';
 export const StyledPopOverMenu = styled.ul`
   width: fit-content;
   max-width: 22rem;
-  ${ellipsis};
   background-color: ${(props) => props.theme.popOver.menuBgd};
   border: ${(props) => props.theme.popOver.menuBorder};
   margin: 0;
@@ -17,6 +16,8 @@ export const StyledPopOverMenu = styled.ul`
   padding: 0.5rem 0;
   box-shadow: ${(props) => props.theme.popOver.shadow};
   z-index: ${(props) => props.theme.zIndex.popOver};
+  display: table;
+  overflow: inherit;
 `;
 
 export const StyledPopOverToggle = styled.button`